+++ toys-core/lib/toys/loader.rb
@@ -86,7 +86,7 @@ module Toys
     # handle the command, loading it from the configuration if necessary, and
     # following aliases.
     # This always returns a tool. If the specific tool path is not defined and
-    # cannot be found in any configuration, it finds the nearest group that
+    # cannot be found in any configuration, it finds the nearest namespace that
     # _would_ contain that tool, up to the root tool.
     #
     # Returns a tuple of the found tool, and the array of remaining arguments