+++ toys-core/test/lookup-cases/normal-file-hierarchy/namespace-1/tool-1-3.rb
@@ -1,6 +1,6 @@
 desc "normal tool-1-3 short description"
 long_desc "normal tool-1-3 long description"
 
-execute do
+script do
   puts "normal tool-1-3 execution"
 end